code-quality: move more utils to utils directory
