+++ packages/cli/src/utils/index.ts
@@ -17,3 +17,5 @@ export { default as delay } from './delay';
 export { fetchDictionary } from './fetchCookieDictionary';
 export { default as validateArgs } from './validateArgs';
 export { default as getUrlListFromArgs } from './getUrlListFromArgs';
+export { default as generateCSVFiles } from './generateCSVfiles';
+export { default as saveCSVReports } from './saveCSVReports';
